requirements.txt
google-genai
cohere
openai
chromadb
django
python-dotenv
Pillow
requests
docling
djangorestframework
reportlab
litellm
PyYAML
django-cors-headers
sentence-transformers
psycopg2-binary
pgvector
dj-database-url
open-clip-torch
numpy
svglib
python-magic
